requirements.txt
flask
waitress
requests
langchain-openai
python-dotenv
PyYAML
hypothesis
pytest
